Setup trend sensors in binary sensor setup
